+++ src/test/testData/inspection/BankCardInString.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class BankCardInString {
+
+    public String card() {
+        return "4539578763621486";
+    }
+}