+++ src/test/testData/inspection/EmailInString.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class EmailInString {
+
+    public String mail() {
+        return "zhangsan@example.com";
+    }
+}