+++ src/test/testData/inspection/EscapedString.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class EscapedString {
+
+    public String value() {
+        return "备注\"张三\" 证件 110101199003078531";
+    }
+}