+++ src/test/testData/inspection/IdCardInComment.java
@@ -1,0 +1,11 @@
+package sample;
+
+/**
+ * 客户张三的证件号是 110101199003078531，请勿外传。
+ */
+public class IdCardInComment {
+
+    /** 联系人证件 110101199003078531 */
+    public void contact() {
+    }
+}