+++ src/test/testData/inspection/InvalidChecksumWithoutContext.java
@@ -1,0 +1,9 @@
+package sample;
+
+public class InvalidChecksumWithoutContext {
+
+    public void run() {
+        String orderNo = "110101199003078532";
+        System.out.println(orderNo);
+    }
+}