+++ src/test/testData/inspection/MockIdCardByFieldName.java
@@ -1,0 +1,10 @@
+package sample;
+
+public class MockIdCardByFieldName {
+
+    private String idCard = "110101199003078532";
+
+    public String getIdCard() {
+        return idCard;
+    }
+}