+++ src/test/testData/inspection/NoSensitiveData.java
@@ -1,0 +1,13 @@
+package sample;
+
+public class NoSensitiveData {
+
+    public String greeting() {
+        return "hello world";
+    }
+
+    // 这是一段没有任何敏感数据的注释
+    public int one() {
+        return 1;
+    }
+}