+++ src/test/testData/quickfix/AfterBothMasked.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class BeforeTwoValues {
+
+    public String value() {
+        return "110101********8531,138****5678";
+    }
+}