+++ src/test/testData/quickfix/AfterIdCard.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class BeforeIdCard {
+
+    public String value() {
+        return "110101********8531";
+    }
+}