+++ src/test/testData/quickfix/BeforeBankCard.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class BeforeBankCard {
+
+    public String value() {
+        return "4539578763621486";
+    }
+}