+++ src/test/testData/quickfix/BeforeEscaped.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class BeforeEscaped {
+
+    public String value() {
+        return "备注\"张三\" 证件 110101199003078531 结束";
+    }
+}