+++ src/test/testData/quickfix/BeforeTwoValues.java
@@ -1,0 +1,8 @@
+package sample;
+
+public class BeforeTwoValues {
+
+    public String value() {
+        return "110101199003078531,13812345678";
+    }
+}